requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.6
django-filter==1.1.0
django-location-field==2.0.3
django-rest-framework-redocs==0.0.1
django-rest-swagger==2.2.0
djangorestframework==3.8.2
GDAL==2.2.3
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
openapi-codec==1.3.2
Pillow==5.1.0
psycopg2==2.7.4
pytz==2018.4
requests==2.19.1
simplejson==3.15.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
